Reorganize testing code.
