+++ tests/simulator/lifecycle_actions.rs
