+++ tests/simulator/token_viewer.rs
